requirements.txt
qrcode[pil]==8.0
reportlab==4.2.5
Pillow==11.1.0
